4.27 Solution
A multithreaded java program that generates the Fibonacci sequence. On the command line, the user will enter the number of Fibonacci numbers that the program is to generate. The program will then create a separate thread that will generate the Fibonacci numbers, placing the sequence in data that can be shared by the threads. The parent thread waits for the child thread to finish, when the child thread finishes execution, the parent thread will output the sequence generated by the child thread.